// api/generated/model/planDayVmDto.ts
/**
 * Generated by orval v8.2.0 🍺
 * Do not edit manually.
 * LgymApi.Api
 * OpenAPI spec version: 1.0
 */
import type { PlanDayExerciseVmDto } from './planDayExerciseVmDto';

export interface PlanDayVmDto {
  /** @nullable */
  _id?: string | null;
  /** @nullable */
  name?: string | null;
  /** @nullable */
  exercises?: PlanDayExerciseVmDto[] | null;
}